+++ cart-service/pom.xml
@@ -37,6 +37,16 @@
             <artifactId>jedis</artifactId>
             <version>3.6.0</version>
         </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-openfeign</artifactId>
+            <version>3.0.4</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.kafka</groupId>
+            <artifactId>spring-kafka</artifactId>
+            <version>2.9.0</version>
+        </dependency>
     </dependencies>
 
 </project>
