+++ cart-service/src/main/java/ru/gb/spring_test/cart/CartApp.java
@@ -3,9 +3,13 @@ package ru.gb.spring_test.cart;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.kafka.annotation.EnableKafka;
 
 @SpringBootApplication
 @EnableCaching
+@EnableFeignClients
+@EnableKafka
 public class CartApp {
     public static void main(String[] args) {
         SpringApplication.run(CartApp.class, args);
