+++ cart-service/src/main/java/ru/gb/spring_test/cart/dto/OrderDetailsDto.java
@@ -1,10 +1,12 @@
 package ru.gb.spring_test.cart.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
+@AllArgsConstructor
 public class OrderDetailsDto {
     private String phone;
     private String address;
