+++ cart-service/src/main/java/ru/gb/spring_test/cart/dto/OrderItemDto.java
@@ -9,14 +9,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class OrderItemDto {
     private Long productId;
-    private String title;
     private int quantity;
     private int pricePerProduct;
     private int price;
 
     public OrderItemDto(ProductDto product) {
         this.productId = product.getId();
-        this.title = product.getTitle();
         this.quantity = 1;
         this.pricePerProduct = product.getCost();
         this.price = product.getCost();
