+++ cart-service/src/main/java/ru/gb/spring_test/cart/services/CartService.java
@@ -1,13 +1,20 @@
 package ru.gb.spring_test.cart.services;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.support.SendResult;
 import org.springframework.stereotype.Service;
+import org.springframework.util.concurrent.ListenableFuture;
 import org.springframework.web.client.RestTemplate;
+import ru.gb.spring_test.cart.api.ProductApi;
+import ru.gb.spring_test.cart.dto.OrderDetailsDto;
+import ru.gb.spring_test.cart.dto.OrderDto;
 import ru.gb.spring_test.cart.dto.ProductDto;
 import ru.gb.spring_test.cart.entities.Cart;
 
@@ -20,7 +27,11 @@ public class CartService {
     @Qualifier("test")
     private final CacheManager cacheManager;
 
-    private final RestTemplate restTemplate;
+    @Qualifier(value = "Kafka")
+    @Autowired
+    private KafkaTemplate<Long, OrderDto> kafkaTemplate;
+
+    private final ProductApi productApi;
 
     @Value("${spring.cache.user.name}")
     private String CACHE_CART;
@@ -41,7 +52,7 @@ public class CartService {
     public Cart addProductByIdToCart(Long id, String cartName) {
         Cart cart = getCurrentCart(cartName);
         if (!cart.addProductCount(id)) {
-            ProductDto product = restTemplate.getForObject("http://localhost:8189/web-market-core/api/v1/products/" + id, ProductDto.class);
+            ProductDto product = productApi.getById(id);
             cart.addProduct(product);
         }
         return cart;
@@ -66,5 +77,20 @@ public class CartService {
         Cart cart = getCurrentCart(cartName);
         cart.decreaseProduct(id);
         return cart;
+    }
+
+    public void createOrder(String username, OrderDetailsDto orderDetailsDto, String cartName) {
+        Cart currentCart = getCurrentCart(cartName);
+        OrderDto order = new OrderDto();
+        order.setAddress(orderDetailsDto.getAddress());
+        order.setPhone(orderDetailsDto.getPhone());
+        order.setUsername(username);
+        order.setTotalPrice(currentCart.getTotalPrice());
+        order.setItemDtoList(currentCart.getItems());
+        clear(cartName);
+//        currentCart.clear();
+        ListenableFuture<SendResult<Long, OrderDto>> future = kafkaTemplate.send("OrdersTopic", order);
+        future.addCallback(System.out::println, System.err::println);
+        kafkaTemplate.flush();
     }
 }
