+++ core-service/pom.xml
@@ -59,6 +59,12 @@
             <version>2.9.0</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-openfeign</artifactId>
+            <version>3.0.4</version>
+        </dependency>
+
     </dependencies>
 
 </project>
