+++ core-service/src/main/java/ru/gb/spring_test/SpringTestApplication.java
@@ -2,10 +2,12 @@ package ru.gb.spring_test;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.kafka.annotation.EnableKafka;
 
 @SpringBootApplication
 @EnableKafka
+@EnableFeignClients
 public class SpringTestApplication {
 
     public static void main(String[] args) {
