+++ core-service/src/main/java/ru/gb/spring_test/configs/KafkaConfig.java
@@ -1,78 +1,78 @@
 package ru.gb.spring_test.configs;
 
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.LongDeserializer;
-import org.apache.kafka.common.serialization.LongSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
-import org.springframework.kafka.config.KafkaListenerContainerFactory;
-import org.springframework.kafka.core.*;
-import org.springframework.kafka.support.serializer.JsonDeserializer;
-import org.springframework.kafka.support.serializer.JsonSerializer;
-import ru.gb.spring_test.dto.FullOrderDetailsDto;
+//import org.apache.kafka.clients.consumer.ConsumerConfig;
+//import org.apache.kafka.clients.producer.ProducerConfig;
+//import org.apache.kafka.common.serialization.LongDeserializer;
+//import org.apache.kafka.common.serialization.LongSerializer;
+//import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+//import org.springframework.kafka.config.KafkaListenerContainerFactory;
+//import org.springframework.kafka.core.*;
+//import org.springframework.kafka.support.serializer.JsonDeserializer;
+//import org.springframework.kafka.support.serializer.JsonSerializer;
+//import ru.gb.spring_test.dto.FullOrderDetailsDto;
+//
+//import java.util.HashMap;
+//import java.util.Map;
 
-import java.util.HashMap;
-import java.util.Map;
-
-@Configuration
-public class KafkaConfig {
-
-    @Value("${spring.kafka.consumer.group-id}")
-    private String groupId;
-
-    @Value("${spring.kafka.bootstrap-servers}")
-    private String server;
-
-    @Bean
-    public Map<String, Object> producerConfig(){
-        Map<String, Object> props = new HashMap<>();
-        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, server);
-        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
-        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, LongSerializer.class);
-        return props;
-    }
-
-    @Bean
-    public JsonDeserializer jsonDeserializer(){
-        JsonDeserializer jsonDeserializer = new JsonDeserializer();
-        jsonDeserializer.addTrustedPackages("*");
-        return jsonDeserializer;
-    }
-
-    @Bean
-    public Map<String, Object> consumerConfig(){
-        Map<String, Object> props = new HashMap<>();
-        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, server);
-        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, JsonDeserializer.class);
-        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, LongDeserializer.class);
-        props.put(JsonDeserializer.TRUSTED_PACKAGES, "*");
-        props.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
-        return props;
-    }
-
-    @Bean
-    public ConsumerFactory<Long, FullOrderDetailsDto> consumerFactory(){
-        return new DefaultKafkaConsumerFactory<>(consumerConfig());
-    }
-
-    @Bean
-    public KafkaListenerContainerFactory<?> kafkaListenerContainerFactory(){
-        ConcurrentKafkaListenerContainerFactory<Long, FullOrderDetailsDto> factory =
-                new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(consumerFactory());
-        return factory;
-    }
-
-    @Bean
-    public ProducerFactory<Long, FullOrderDetailsDto> producerFactory(){
-      return new DefaultKafkaProducerFactory<>(producerConfig());
-    }
-
-    @Bean(value = "Kafka")
-    public KafkaTemplate<Long, FullOrderDetailsDto> kafkaTemplate(){
-        return new KafkaTemplate<>(producerFactory());
-    }
-}
+//@Configuration
+//public class KafkaConfig {
+//
+//    @Value("${spring.kafka.consumer.group-id}")
+//    private String groupId;
+//
+//    @Value("${spring.kafka.bootstrap-servers}")
+//    private String server;
+//
+//    @Bean
+//    public Map<String, Object> producerConfig(){
+//        Map<String, Object> props = new HashMap<>();
+//        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, server);
+//        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
+//        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, LongSerializer.class);
+//        return props;
+//    }
+//
+//    @Bean
+//    public JsonDeserializer jsonDeserializer(){
+//        JsonDeserializer jsonDeserializer = new JsonDeserializer();
+//        jsonDeserializer.addTrustedPackages("*");
+//        return jsonDeserializer;
+//    }
+//
+//    @Bean
+//    public Map<String, Object> consumerConfig(){
+//        Map<String, Object> props = new HashMap<>();
+//        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, server);
+//        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, JsonDeserializer.class);
+//        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, LongDeserializer.class);
+//        props.put(JsonDeserializer.TRUSTED_PACKAGES, "*");
+//        props.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
+//        return props;
+//    }
+//
+//    @Bean
+//    public ConsumerFactory<Long, FullOrderDetailsDto> consumerFactory(){
+//        return new DefaultKafkaConsumerFactory<>(consumerConfig());
+//    }
+//
+//    @Bean
+//    public KafkaListenerContainerFactory<?> kafkaListenerContainerFactory(){
+//        ConcurrentKafkaListenerContainerFactory<Long, FullOrderDetailsDto> factory =
+//                new ConcurrentKafkaListenerContainerFactory<>();
+//        factory.setConsumerFactory(consumerFactory());
+//        return factory;
+//    }
+//
+//    @Bean
+//    public ProducerFactory<Long, FullOrderDetailsDto> producerFactory(){
+//      return new DefaultKafkaProducerFactory<>(producerConfig());
+//    }
+//
+//    @Bean(value = "Kafka")
+//    public KafkaTemplate<Long, FullOrderDetailsDto> kafkaTemplate(){
+//        return new KafkaTemplate<>(producerFactory());
+//    }
+//}
