+++ core-service/src/main/java/ru/gb/spring_test/converters/OrderConverter.java
@@ -1,29 +1,29 @@
 package ru.gb.spring_test.converters;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import ru.gb.spring_test.dto.OrderDto;
-import ru.gb.spring_test.entities.Order;
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.stereotype.Component;
+//import ru.gb.spring_test.dto.OrderDto;
+//import ru.gb.spring_test.entities.Order;
+//
+//import java.util.stream.Collectors;
 
-import java.util.stream.Collectors;
-
-@Component
-@RequiredArgsConstructor
-public class OrderConverter {
-    private final OrderItemConverter orderItemConverter;
-
-    public Order dtoToEntity(OrderDto orderDto) {
-        throw new UnsupportedOperationException();
-    }
-
-    public OrderDto entityToDto(Order order) {
-        OrderDto out = new OrderDto();
-        out.setId(order.getId());
-        out.setAddress(order.getAddress());
-        out.setPhone(order.getPhone());
-        out.setTotalPrice(order.getTotalPrice());
-        out.setUsername(order.getAddress());
-        out.setItemDtoList(order.getItems().stream().map(orderItemConverter::entityToDto).collect(Collectors.toList()));
-        return out;
-    }
-}
+//@Component
+//@RequiredArgsConstructor
+//public class OrderConverter {
+//    private final OrderItemConverter orderItemConverter;
+//
+//    public Order dtoToEntity(OrderDto orderDto) {
+//        throw new UnsupportedOperationException();
+//    }
+//
+//    public OrderDto entityToDto(Order order) {
+//        OrderDto out = new OrderDto();
+//        out.setId(order.getId());
+//        out.setAddress(order.getAddress());
+//        out.setPhone(order.getPhone());
+//        out.setTotalPrice(order.getTotalPrice());
+//        out.setUsername(order.getAddress());
+//        out.setItemDtoList(order.getItems().stream().map(orderItemConverter::entityToDto).collect(Collectors.toList()));
+//        return out;
+//    }
+//}
