+++ core-service/src/main/java/ru/gb/spring_test/dto/FullOrderDetailsDto.java
@@ -1,15 +1,15 @@
 package ru.gb.spring_test.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+//import lombok.AllArgsConstructor;
+//import lombok.Data;
+//import lombok.NoArgsConstructor;
 
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class FullOrderDetailsDto {
-
-    private String username;
-    private OrderDetailsDto orderDetailsDto;
-    private String cartName;
-}
+//@Data
+//@AllArgsConstructor
+//@NoArgsConstructor
+//public class FullOrderDetailsDto {
+//
+//    private String username;
+//    private OrderDetailsDto orderDetailsDto;
+//    private String cartName;
+//}
