+++ core-service/src/main/java/ru/gb/spring_test/dto/OrderDetailsDto.java
@@ -1,13 +1,13 @@
 package ru.gb.spring_test.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+//import lombok.AllArgsConstructor;
+//import lombok.Data;
+//import lombok.NoArgsConstructor;
 
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class OrderDetailsDto {
-    private String phone;
-    private String address;
-}
+//@Data
+//@NoArgsConstructor
+//@AllArgsConstructor
+//public class OrderDetailsDto {
+//    private String phone;
+//    private String address;
+//}
