+++ core-service/src/main/java/ru/gb/spring_test/dto/OrderItemDto.java
@@ -1,30 +1,30 @@
 package ru.gb.spring_test.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import ru.gb.spring_test.entities.Product;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class OrderItemDto {
-    private Long productId;
-    private String title;
-    private int quantity;
-    private int pricePerProduct;
-    private int price;
-
-    public OrderItemDto(Product product) {
-        this.productId = product.getId();
-        this.title = product.getTitle();
-        this.quantity = 1;
-        this.pricePerProduct = product.getCost();
-        this.price = product.getCost();
-    }
-
-    public void changeQuantity(int delta) {
-        this.quantity += delta;
-        this.price = this.quantity * this.pricePerProduct;
-    }
-}
+//import lombok.AllArgsConstructor;
+//import lombok.Data;
+//import lombok.NoArgsConstructor;
+//import ru.gb.spring_test.entities.Product;
+//
+//@Data
+//@AllArgsConstructor
+//@NoArgsConstructor
+//public class OrderItemDto {
+//    private Long productId;
+//    private String title;
+//    private int quantity;
+//    private int pricePerProduct;
+//    private int price;
+//
+//    public OrderItemDto(Product product) {
+//        this.productId = product.getId();
+//        this.title = product.getTitle();
+//        this.quantity = 1;
+//        this.pricePerProduct = product.getCost();
+//        this.price = product.getCost();
+//    }
+//
+//    public void changeQuantity(int delta) {
+//        this.quantity += delta;
+//        this.price = this.quantity * this.pricePerProduct;
+//    }
+//}
