+++ core-service/src/main/java/ru/gb/spring_test/entities/Order.java
@@ -1,41 +1,41 @@
 package ru.gb.spring_test.entities;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
+//import lombok.AllArgsConstructor;
+//import lombok.Data;
+//import lombok.NoArgsConstructor;
+//import org.hibernate.annotations.CreationTimestamp;
+//import org.hibernate.annotations.UpdateTimestamp;
+//
+//import javax.persistence.*;
+//import java.time.LocalDateTime;
+//import java.util.List;
 
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Data
-@Entity
-@Table(name = "orders")
-@NoArgsConstructor
-@AllArgsConstructor
-public class Order {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String username;
-
-    @OneToMany(mappedBy = "order", cascade = {CascadeType.PERSIST, CascadeType.REMOVE})
-    private List<OrderItem> items;
-
-    private Integer totalPrice;
-
-    private String address;
-
-    private String phone;
-
-    @CreationTimestamp
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    private LocalDateTime updatedAt;
-
-}
+//@Data
+//@Entity
+//@Table(name = "orders")
+//@NoArgsConstructor
+//@AllArgsConstructor
+//public class Order {
+//
+//    @Id
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    private Long id;
+//
+//    private String username;
+//
+//    @OneToMany(mappedBy = "order", cascade = {CascadeType.PERSIST, CascadeType.REMOVE})
+//    private List<OrderItem> items;
+//
+//    private Integer totalPrice;
+//
+//    private String address;
+//
+//    private String phone;
+//
+//    @CreationTimestamp
+//    private LocalDateTime createdAt;
+//
+//    @UpdateTimestamp
+//    private LocalDateTime updatedAt;
+//
+//}
