+++ core-service/src/main/java/ru/gb/spring_test/entities/OrderItem.java
@@ -1,42 +1,42 @@
 package ru.gb.spring_test.entities;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
+//import lombok.AllArgsConstructor;
+//import lombok.Data;
+//import lombok.NoArgsConstructor;
+//import org.hibernate.annotations.CreationTimestamp;
+//import org.hibernate.annotations.UpdateTimestamp;
+//
+//import javax.persistence.*;
+//import java.time.LocalDateTime;
 
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "order_items")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class OrderItem {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    @CreationTimestamp
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    private LocalDateTime updatedAt;
-
-    private Integer quantity;
-
-    private Integer pricePerProduct;
-
-    private Integer price;
-
-    @ManyToOne
-    @JoinColumn(name = "order_id")
-    private Order order;
-}
+//@Entity
+//@Table(name = "order_items")
+//@Data
+//@NoArgsConstructor
+//@AllArgsConstructor
+//public class OrderItem {
+//
+//    @Id
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    private Long id;
+//
+//    @ManyToOne
+//    @JoinColumn(name = "product_id")
+//    private Product product;
+//
+//    @CreationTimestamp
+//    private LocalDateTime createdAt;
+//
+//    @UpdateTimestamp
+//    private LocalDateTime updatedAt;
+//
+//    private Integer quantity;
+//
+//    private Integer pricePerProduct;
+//
+//    private Integer price;
+//
+//    @ManyToOne
+//    @JoinColumn(name = "order_id")
+//    private Order order;
+//}
