+++ core-service/src/main/java/ru/gb/spring_test/repositories/OrderRepository.java
@@ -1,15 +1,15 @@
 package ru.gb.spring_test.repositories;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-import ru.gb.spring_test.entities.Order;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//import org.springframework.data.jpa.repository.Query;
+//import org.springframework.stereotype.Repository;
+//import ru.gb.spring_test.entities.Order;
+//
+//import java.util.List;
 
-import java.util.List;
-
-@Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {
-
-    @Query("select o from Order o where o.username = :username")
-    List<Order> findByUsername(String username);
-}
+//@Repository
+//public interface OrderRepository extends JpaRepository<Order, Long> {
+//
+//    @Query("select o from Order o where o.username = :username")
+//    List<Order> findByUsername(String username);
+//}
