+++ order-service/src/main/java/ru/gb/spring_test/order/OrderApp.java
@@ -1,0 +1,17 @@
+package ru.gb.spring_test.order;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.kafka.annotation.EnableKafka;
+
+@SpringBootApplication
+@EnableKafka
+@EnableFeignClients
+public class OrderApp {
+
+    public static void main(String[] args) {
+        SpringApplication.run(OrderApp.class, args);
+    }
+
+}
