+++ order-service/src/main/java/ru/gb/spring_test/order/configs/KafkaConfig.java
@@ -1,0 +1,58 @@
+package ru.gb.spring_test.order.configs;
+
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.annotation.EnableKafka;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.config.KafkaListenerContainerFactory;
+import org.springframework.kafka.core.*;
+import org.springframework.kafka.support.serializer.JsonDeserializer;
+import ru.gb.spring_test.order.dto.OrderDto;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Configuration
+@EnableKafka
+public class KafkaConfig {
+
+    @Value("${spring.kafka.consumer.group-id}")
+    private String groupId;
+
+    @Value("${spring.kafka.bootstrap-servers}")
+    private String server;
+
+    @Bean
+    public JsonDeserializer jsonDeserializer(){
+        JsonDeserializer jsonDeserializer = new JsonDeserializer();
+        jsonDeserializer.addTrustedPackages("*");
+        return jsonDeserializer;
+    }
+
+    @Bean
+    public Map<String, Object> consumerConfig(){
+        Map<String, Object> props = new HashMap<>();
+        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, server);
+        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, JsonDeserializer.class);
+        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, JsonDeserializer.class);
+        props.put(JsonDeserializer.TRUSTED_PACKAGES, "*");
+        props.put(JsonDeserializer.TYPE_MAPPINGS, "OrderDto:ru.gb.spring_test.order.dto.OrderDto");
+        props.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
+        return props;
+    }
+
+    @Bean
+    public ConsumerFactory<Long, OrderDto> consumerFactory(){
+        return new DefaultKafkaConsumerFactory<>(consumerConfig());
+    }
+
+    @Bean
+    public KafkaListenerContainerFactory<?> kafkaListenerContainerFactory(){
+        ConcurrentKafkaListenerContainerFactory<Long, OrderDto> factory =
+                new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setConsumerFactory(consumerFactory());
+        return factory;
+    }
+}
