+++ order-service/src/main/java/ru/gb/spring_test/order/controllers/OrderController.java
@@ -1,0 +1,25 @@
+package ru.gb.spring_test.order.controllers;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+import ru.gb.spring_test.order.converters.OrderConverter;
+import ru.gb.spring_test.order.dto.OrderDto;
+import ru.gb.spring_test.order.services.OrderService;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@RestController
+@RequestMapping("/api/v1/orders")
+@RequiredArgsConstructor
+public class OrderController {
+
+    private final OrderService orderService;
+    private final OrderConverter orderConverter;
+
+    @GetMapping
+    public List<OrderDto> getCurrenOrders(@RequestHeader String username){
+        return orderService.findOrdersByUsername(username).stream()
+                .map(orderConverter::entityToDto).collect(Collectors.toList());
+    }
+}
