+++ order-service/src/main/java/ru/gb/spring_test/order/converters/OrderConverter.java
@@ -1,0 +1,29 @@
+package ru.gb.spring_test.order.converters;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import ru.gb.spring_test.order.dto.OrderDto;
+import ru.gb.spring_test.order.entities.Order;
+
+import java.util.stream.Collectors;
+
+@Component
+@RequiredArgsConstructor
+public class OrderConverter {
+    private final OrderItemConverter orderItemConverter;
+
+    public Order dtoToEntity(OrderDto orderDto) {
+        throw new UnsupportedOperationException();
+    }
+
+    public OrderDto entityToDto(Order order) {
+        OrderDto out = new OrderDto();
+        out.setId(order.getId());
+        out.setAddress(order.getAddress());
+        out.setPhone(order.getPhone());
+        out.setTotalPrice(order.getTotalPrice());
+        out.setUsername(order.getAddress());
+        out.setItemDtoList(order.getItems().stream().map(orderItemConverter::entityToDto).collect(Collectors.toList()));
+        return out;
+    }
+}
