+++ order-service/src/main/java/ru/gb/spring_test/order/dto/Cart.java
@@ -1,0 +1,74 @@
+package ru.gb.spring_test.order.dto;
+
+import lombok.Data;
+import org.springframework.cache.CacheManager;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+
+@Data
+public class Cart {
+    private List<OrderItemDto> items;
+    private int totalPrice;
+
+    public Cart() {
+        this.items = new ArrayList<>();
+    }
+
+    public Cart(String cartName, CacheManager manager){
+        this.items = new ArrayList<>();
+        this.totalPrice = 0;
+    }
+
+    public boolean addProductCount(Long id) {
+        for (OrderItemDto o: items) {
+            if (o.getProductId().equals(id)) {
+                o.changeQuantity(1);
+                recalculate();
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public void addProduct(ProductDto product) {
+        if (addProductCount(product.getId())) {
+            return;
+        }
+        items.add(new OrderItemDto(product));
+        recalculate();
+    }
+
+    private void recalculate() {
+        totalPrice = 0;
+        for (OrderItemDto o: items) {
+            totalPrice += o.getPrice();
+        }
+    }
+
+    public void clear() {
+        items.clear();
+        totalPrice = 0;
+    }
+
+    public void removeProduct(Long id) {
+        items.removeIf(o -> o.getProductId().equals(id));
+        recalculate();
+    }
+
+    public void decreaseProduct(Long id) {
+        Iterator<OrderItemDto> iter = items.iterator();
+        while (iter.hasNext()) {
+            OrderItemDto o = iter.next();
+            if (o.getProductId().equals(id)) {
+                o.changeQuantity(-1);
+                if (o.getQuantity() <= 0) {
+                    iter.remove();
+                }
+                recalculate();
+                return;
+            }
+        }
+    }
+}
