+++ order-service/src/main/java/ru/gb/spring_test/order/dto/OrderDetailsDto.java
@@ -1,0 +1,13 @@
+package ru.gb.spring_test.order.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class OrderDetailsDto {
+    private String phone;
+    private String address;
+}
