+++ order-service/src/main/java/ru/gb/spring_test/order/dto/OrderDto.java
@@ -1,0 +1,16 @@
+package ru.gb.spring_test.order.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class OrderDto {
+
+    private Long id;
+    private String username;
+    private List<OrderItemDto> itemDtoList;
+    private Integer totalPrice;
+    private String address;
+    private String phone;
+}
