+++ order-service/src/main/java/ru/gb/spring_test/order/dto/OrderItemDto.java
@@ -1,0 +1,27 @@
+package ru.gb.spring_test.order.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class OrderItemDto {
+    private Long productId;
+    private int quantity;
+    private int pricePerProduct;
+    private int price;
+
+    public OrderItemDto(ProductDto product) {
+        this.productId = product.getId();
+        this.quantity = 1;
+        this.pricePerProduct = product.getCost();
+        this.price = product.getCost();
+    }
+
+    public void changeQuantity(int delta) {
+        this.quantity += delta;
+        this.price = this.quantity * this.pricePerProduct;
+    }
+}
