+++ order-service/src/main/java/ru/gb/spring_test/order/dto/ProductDto.java
@@ -1,0 +1,18 @@
+package ru.gb.spring_test.order.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ProductDto {
+
+    private Long id;
+    private String title;
+    private Integer cost;
+
+}
