+++ order-service/src/main/java/ru/gb/spring_test/order/exceptions/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package ru.gb.spring_test.order.exceptions;
+
+public class ResourceNotFoundException extends RuntimeException{
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
