+++ order-service/src/main/java/ru/gb/spring_test/order/repositories/OrderRepository.java
@@ -1,0 +1,15 @@
+package ru.gb.spring_test.order.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+import ru.gb.spring_test.order.entities.Order;
+
+import java.util.List;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
+    @Query("select o from Order o where o.username = :username")
+    List<Order> findByUsername(String username);
+}
