+++ order-service/src/main/java/ru/gb/spring_test/order/services/OrderService.java
@@ -1,0 +1,51 @@
+package ru.gb.spring_test.order.services;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import ru.gb.spring_test.order.dto.OrderDto;
+import ru.gb.spring_test.order.entities.Order;
+import ru.gb.spring_test.order.entities.OrderItem;
+import ru.gb.spring_test.order.repositories.OrderRepository;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+public class OrderService {
+
+    private final OrderRepository orderRepository;
+
+    @Transactional
+    @KafkaListener(topics = "OrdersTopic")
+    public void saveOrder(OrderDto orderDto){
+        Order order = new Order();
+        order.setAddress(orderDto.getAddress());
+        order.setPhone(orderDto.getPhone());
+        order.setUsername(orderDto.getUsername());
+        order.setTotalPrice(orderDto.getTotalPrice());
+        List<OrderItem> items = orderDto.getItemDtoList().stream()
+                .map(o -> {
+                    OrderItem orderItem = new OrderItem();
+                    orderItem.setOrder(order);
+                    orderItem.setQuantity(o.getQuantity());
+                    orderItem.setPricePerProduct(o.getPricePerProduct());
+                    orderItem.setPrice(o.getPrice());
+                    orderItem.setProductId(o.getProductId());
+                    return orderItem;
+                }).collect(Collectors.toList());
+        order.setItems(items);
+        orderRepository.save(order);
+    }
+
+    public List<Order> findOrdersByUsername(String username) {
+        try {
+            return orderRepository.findByUsername(username);
+        } catch (Exception e) {
+            return new ArrayList<>();
+        }
+    }
+}
