+++ order-service/src/main/resources/application.yaml
@@ -1,0 +1,19 @@
+server:
+  port: 8186
+  servlet:
+    context-path: /web-market-order
+spring:
+  datasource:
+    driver-class-name: org.postgresql.Driver
+    url: jdbc:postgresql://localhost:5432/productsDB
+    username: postgres
+    password: 739285
+  jpa:
+    show-sql: true
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQL94Dialect
+  kafka:
+    consumer:
+      group-id: 123
+    bootstrap-servers: localhost:9092
